requirements.txt
ansicolors==1.1.8
appdirs==1.4.3
#apturl==0.5.2
asgiref==3.3.4
bcrypt==3.1.7
black==21.7b0
blinker==1.4
#Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
#chrome-gnome-shell==0.0.0
#cliapp==1.20180812.1
click==8.0.1
cm-rgb==0.3.6
#cmdtest==0.32+git
codecov==2.1.12
collective.checkdocs==0.2
colorama==0.4.3
#command-not-found==0.3
coverage==5.5
crayons==0.4.0
cryptography==2.8
#cupshelpers==1.0
dbus-python==1.2.16
#defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
#distro==1.4.0
#distro-info===0.23ubuntu1

django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-excel==0.0.10
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
djangorestframework==3.12.4
docutils==0.17.1
#duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.0.12
flake8==3.9.2
fs==2.4.13
future==0.18.2
gease==0.0.11
gunicorn==20.1.0
hidapi==0.10.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.9.3
Jinja2==3.0.1
jinja2-fsloader==0.3.0
keyring==18.0.1
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lml==0.1.0
lockfile==0.12.2
#louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
moban==0.8.2
moban-jinja2-github==0.0.4
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
#mysqlclient==1.4.4
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.7
packaging==20.3
paramiko==2.6.0
pathspec==0.9.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.8.0
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
pyecharts-jupyter-installer==0.0.3
pyexcel==0.6.6
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.2
pyexcel-io==0.6.4
pyexcel-ods==0.6.0
pyexcel-ods3==0.6.0
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pyflakes==2.3.1
pygal==2.4.0
Pygments==2.3.1
#PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyRFC3339==1.1
#python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
#python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.8.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
ssh-import-id==5.10
supervisor==4.1.0
#systemd-python==234
tablib==3.0.0
texttable==1.6.4
tomli==1.2.1
#ttystatus==0.38
twilio==6.63.0
#ubuntu-advantage-tools==27.2
#ubuntu-drivers-common==0.0.0
#ufw==0.36
#unattended-upgrades==0.1
urllib3==1.25.8
#usb-creator==0.3.7
#vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==5.2.0
#xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
#zipp==1.0.0
